pyproject.toml
[project]
name = "smart-season-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.9",
    "djoser>=2.3.3",
    "drf-spectacular>=0.29.0",
    "gunicorn>=25.3.0",
    "psycopg2-binary>=2.9.11",
    "whitenoise>=6.12.0",
]
